+++ Assets/Group1/Scripts/SpeedBoost.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class SpeedBoost : MoveIncideCircle
+public class SpeedBoost : MonoBehaviour
 {
     [SerializeField] float _speedMultiplier = 2;
     [SerializeField] float _boostDuration = 3;
